Verify that styles are loaded correctly.
