+++ robolectric-resources/src/main/java/org/robolectric/res/PluralResourceLoader.java
@@ -21,19 +21,4 @@ public class PluralResourceLoader extends XpathResourceXmlLoader {
     resourceTable.addResource("plurals", name, new PluralRules(rules, ResType.CHAR_SEQUENCE, xmlContext));
   }
 
-  public static class PluralRules extends TypedResource<List<Plural>> {
-    public PluralRules(List<Plural> data, ResType resType, XmlContext xmlContext) {
-      super(data, resType, xmlContext);
-    }
-
-    public Plural find(int quantity) {
-      for (Plural p : getData()) {
-        if (p.num == quantity) return p;
-      }
-      for (Plural p : getData()) {
-        if (p.num == -1) return p;
-      }
-      return null;
-    }
-  }
 }
