+++ robolectric-resources/src/main/java/org/robolectric/res/PluralRules.java
@@ -1,0 +1,19 @@
+package org.robolectric.res;
+
+import java.util.List;
+
+public class PluralRules extends TypedResource<List<Plural>> {
+  public PluralRules(List<Plural> data, ResType resType, XmlContext xmlContext) {
+    super(data, resType, xmlContext);
+  }
+
+  public Plural find(int quantity) {
+    for (Plural p : getData()) {
+      if (p.num == quantity) return p;
+    }
+    for (Plural p : getData()) {
+      if (p.num == -1) return p;
+    }
+    return null;
+  }
+}
